Extract MigrationGenerator from Migrator
